refactor(api):[#943] refactor according to feedback
