+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/assets/application/asbuilt/rest/AssetAsBuiltController.java
@@ -22,6 +22,7 @@
 package org.eclipse.tractusx.traceability.assets.application.asbuilt.rest;
 
 import assets.importpoc.ErrorResponse;
+import assets.request.SearchableAssetsRequest;
 import assets.response.asbuilt.AssetAsBuiltResponse;
 import assets.response.base.request.UpdateAssetRequest;
 import io.swagger.v3.oas.annotations.Operation;
@@ -37,8 +38,8 @@ import org.apache.commons.lang3.ArrayUtils;
 import org.eclipse.tractusx.traceability.assets.application.asbuilt.mapper.AssetAsBuiltFieldMapper;
 import org.eclipse.tractusx.traceability.assets.application.asbuilt.mapper.AssetAsBuiltResponseMapper;
 import org.eclipse.tractusx.traceability.assets.application.asbuilt.mapper.QualityTypeMapper;
+import org.eclipse.tractusx.traceability.assets.application.base.mapper.OwnerTypeMapper;
 import org.eclipse.tractusx.traceability.assets.application.base.request.GetDetailInformationRequest;
-import org.eclipse.tractusx.traceability.assets.application.base.request.SearchableAssetsRequest;
 import org.eclipse.tractusx.traceability.assets.application.base.request.SyncAssetsRequest;
 import org.eclipse.tractusx.traceability.assets.application.base.service.AssetBaseService;
 import org.eclipse.tractusx.traceability.common.model.BaseRequestFieldMapper;
@@ -258,7 +259,7 @@ public class AssetAsBuiltController {
         int size = searchableAssetsRequest.size() == null ? 200 : searchableAssetsRequest.size();
 
         return assetBaseService.getSearchableValues(fieldMapper.mapRequestFieldName(searchableAssetsRequest.fieldName()),
-                searchableAssetsRequest.startWith(), size, searchableAssetsRequest.owner(), inAssetIdsList);
+                searchableAssetsRequest.startWith(), size, OwnerTypeMapper.from(searchableAssetsRequest.owner()), inAssetIdsList);
     }
 
     @Operation(operationId = "assetsCountryMap",
