+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/assets/application/base/mapper/OwnerTypeMapper.java
@@ -1,0 +1,33 @@
+/********************************************************************************
+ * Copyright (c) 2024 Contributors to the Eclipse Foundation
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Apache License, Version 2.0 which is available at
+ * https://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ ********************************************************************************/
+
+package org.eclipse.tractusx.traceability.assets.application.base.mapper;
+
+import assets.response.base.OwnerType;
+import org.eclipse.tractusx.traceability.assets.domain.base.model.Owner;
+
+import static java.util.Objects.nonNull;
+
+public abstract class OwnerTypeMapper {
+    public static Owner from(final OwnerType type) {
+        if (nonNull(type)) {
+            return Owner.valueOf(type.name());
+        } else return null;
+    }
+}
